Module rscam::consts

Constants

AUTO_FOCUS_RANGE_AUTO

AUTO_FOCUS_RANGE_INFINITY

AUTO_FOCUS_RANGE_MACRO

AUTO_FOCUS_RANGE_NORMAL

AUTO_FOCUS_STATUS_BUSY

AUTO_FOCUS_STATUS_FAILED

AUTO_FOCUS_STATUS_IDLE

AUTO_FOCUS_STATUS_REACHED

CID_3A_LOCK

CID_ALPHA_COMPONENT

CID_ANALOGUE_GAIN

CID_AUDIO_BALANCE

CID_AUDIO_BASS

CID_AUDIO_COMPRESSION_ATTACK_TIME

CID_AUDIO_COMPRESSION_ENABLED

CID_AUDIO_COMPRESSION_GAIN

CID_AUDIO_COMPRESSION_RELEASE_TIME

CID_AUDIO_COMPRESSION_THRESHOLD

CID_AUDIO_LIMITER_DEVIATION

CID_AUDIO_LIMITER_ENABLED

CID_AUDIO_LIMITER_RELEASE_TIME

CID_AUDIO_LOUDNESS

CID_AUDIO_MUTE

CID_AUDIO_TREBLE

CID_AUDIO_VOLUME

CID_AUTOBRIGHTNESS

CID_AUTOGAIN

CID_AUTO_EXPOSURE_BIAS

CID_AUTO_FOCUS_RANGE

CID_AUTO_FOCUS_START

CID_AUTO_FOCUS_STATUS

CID_AUTO_FOCUS_STOP

CID_AUTO_N_PRESET_WHITE_BALANCE

CID_AUTO_WHITE_BALANCE

CID_BACKLIGHT_COMPENSATION

CID_BAND_STOP_FILTER

CID_BASE

CID_BG_COLOR

CID_BLACK_LEVEL

CID_BLUE_BALANCE

CID_BRIGHTNESS

CID_CAMERA_CLASS

CID_CAMERA_CLASS_BASE

CID_CHROMA_AGC

CID_CHROMA_GAIN

CID_COLORFX

CID_COLORFX_CBCR

CID_COLOR_KILLER

CID_CONTRAST

CID_DETECT_CLASS

CID_DETECT_CLASS_BASE

CID_DETECT_MD_GLOBAL_THRESHOLD

CID_DETECT_MD_MODE

CID_DETECT_MD_REGION_GRID

CID_DETECT_MD_THRESHOLD_GRID

CID_DO_WHITE_BALANCE

CID_DV_CLASS

CID_DV_CLASS_BASE

CID_DV_RX_POWER_PRESENT

CID_DV_RX_RGB_RANGE

CID_DV_TX_EDID_PRESENT

CID_DV_TX_HOTPLUG

CID_DV_TX_MODE

CID_DV_TX_RGB_RANGE

CID_DV_TX_RXSENSE

CID_EXPOSURE

CID_EXPOSURE_ABSOLUTE

CID_EXPOSURE_AUTO

CID_EXPOSURE_AUTO_PRIORITY

CID_EXPOSURE_METERING

CID_FLASH_CHARGE

CID_FLASH_CLASS

CID_FLASH_CLASS_BASE

CID_FLASH_FAULT

CID_FLASH_INDICATOR_INTENSITY

CID_FLASH_INTENSITY

CID_FLASH_LED_MODE

CID_FLASH_READY

CID_FLASH_STROBE

CID_FLASH_STROBE_SOURCE

CID_FLASH_STROBE_STATUS

CID_FLASH_STROBE_STOP

CID_FLASH_TIMEOUT

CID_FLASH_TORCH_INTENSITY

CID_FM_RX_CLASS

CID_FM_RX_CLASS_BASE

CID_FM_TX_CLASS

CID_FM_TX_CLASS_BASE

CID_FOCUS_ABSOLUTE

CID_FOCUS_AUTO

CID_FOCUS_RELATIVE

CID_GAIN

CID_GAMMA

CID_HBLANK

CID_HFLIP

CID_HUE

CID_HUE_AUTO

CID_ILLUMINATORS_1

CID_ILLUMINATORS_2

CID_IMAGE_PROC_CLASS

CID_IMAGE_PROC_CLASS_BASE

CID_IMAGE_SOURCE_CLASS

CID_IMAGE_SOURCE_CLASS_BASE

CID_IMAGE_STABILIZATION

CID_IRIS_ABSOLUTE

CID_IRIS_RELATIVE

CID_ISO_SENSITIVITY

CID_ISO_SENSITIVITY_AUTO

CID_JPEG_ACTIVE_MARKER

CID_JPEG_CHROMA_SUBSAMPLING

CID_JPEG_CLASS

CID_JPEG_CLASS_BASE

CID_JPEG_COMPRESSION_QUALITY

CID_JPEG_RESTART_INTERVAL

CID_LASTP1

CID_LINK_FREQ

CID_MIN_BUFFERS_FOR_CAPTURE

CID_MIN_BUFFERS_FOR_OUTPUT

CID_MPEG_AUDIO_AAC_BITRATE

CID_MPEG_AUDIO_AC3_BITRATE

CID_MPEG_AUDIO_CRC

CID_MPEG_AUDIO_DEC_MULTILINGUAL_PLAYBACK

CID_MPEG_AUDIO_DEC_PLAYBACK

CID_MPEG_AUDIO_EMPHASIS

CID_MPEG_AUDIO_ENCODING

CID_MPEG_AUDIO_L1_BITRATE

CID_MPEG_AUDIO_L2_BITRATE

CID_MPEG_AUDIO_L3_BITRATE

CID_MPEG_AUDIO_MODE

CID_MPEG_AUDIO_MODE_EXTENSION

CID_MPEG_AUDIO_MUTE

CID_MPEG_AUDIO_SAMPLING_FREQ

CID_MPEG_BASE

CID_MPEG_CLASS

CID_MPEG_CX2341X_BASE

CID_MPEG_CX2341X_STREAM_INSERT_NAV_PACKETS

CID_MPEG_CX2341X_VIDEO_CHROMA_MEDIAN_FILTER_BOTTOM

CID_MPEG_CX2341X_VIDEO_CHROMA_MEDIAN_FILTER_TOP

CID_MPEG_CX2341X_VIDEO_CHROMA_SPATIAL_FILTER_TYPE

CID_MPEG_CX2341X_VIDEO_LUMA_MEDIAN_FILTER_BOTTOM

CID_MPEG_CX2341X_VIDEO_LUMA_MEDIAN_FILTER_TOP

CID_MPEG_CX2341X_VIDEO_LUMA_SPATIAL_FILTER_TYPE

CID_MPEG_CX2341X_VIDEO_MEDIAN_FILTER_TYPE

CID_MPEG_CX2341X_VIDEO_SPATIAL_FILTER

CID_MPEG_CX2341X_VIDEO_SPATIAL_FILTER_MODE

CID_MPEG_CX2341X_VIDEO_TEMPORAL_FILTER

CID_MPEG_CX2341X_VIDEO_TEMPORAL_FILTER_MODE

CID_MPEG_MFC51_BASE

CID_MPEG_MFC51_VIDEO_DECODER_H264_DISPLAY_DELAY

CID_MPEG_MFC51_VIDEO_DECODER_H264_DISPLAY_DELAY_ENABLE

CID_MPEG_MFC51_VIDEO_FORCE_FRAME_TYPE

CID_MPEG_MFC51_VIDEO_FRAME_SKIP_MODE

CID_MPEG_MFC51_VIDEO_H264_ADAPTIVE_RC_ACTIVITY

CID_MPEG_MFC51_VIDEO_H264_ADAPTIVE_RC_DARK

CID_MPEG_MFC51_VIDEO_H264_ADAPTIVE_RC_SMOOTH

CID_MPEG_MFC51_VIDEO_H264_ADAPTIVE_RC_STATIC

CID_MPEG_MFC51_VIDEO_H264_NUM_REF_PIC_FOR_P

CID_MPEG_MFC51_VIDEO_PADDING

CID_MPEG_MFC51_VIDEO_PADDING_YUV

CID_MPEG_MFC51_VIDEO_RC_FIXED_TARGET_BIT

CID_MPEG_MFC51_VIDEO_RC_REACTION_COEFF

CID_MPEG_STREAM_PES_ID_AUDIO

CID_MPEG_STREAM_PES_ID_VIDEO

CID_MPEG_STREAM_PID_AUDIO

CID_MPEG_STREAM_PID_PCR

CID_MPEG_STREAM_PID_PMT

CID_MPEG_STREAM_PID_VIDEO

CID_MPEG_STREAM_TYPE

CID_MPEG_STREAM_VBI_FMT

CID_MPEG_VIDEO_ASPECT

CID_MPEG_VIDEO_BITRATE

CID_MPEG_VIDEO_BITRATE_MODE

CID_MPEG_VIDEO_BITRATE_PEAK

CID_MPEG_VIDEO_B_FRAMES

CID_MPEG_VIDEO_CYCLIC_INTRA_REFRESH_MB

CID_MPEG_VIDEO_DECODER_MPEG4_DEBLOCK_FILTER

CID_MPEG_VIDEO_DECODER_SLICE_INTERFACE

CID_MPEG_VIDEO_DEC_FRAME

CID_MPEG_VIDEO_DEC_PTS

CID_MPEG_VIDEO_ENCODING

CID_MPEG_VIDEO_FRAME_RC_ENABLE

CID_MPEG_VIDEO_GOP_CLOSURE

CID_MPEG_VIDEO_GOP_SIZE

CID_MPEG_VIDEO_H263_B_FRAME_QP

CID_MPEG_VIDEO_H263_I_FRAME_QP

CID_MPEG_VIDEO_H263_MAX_QP

CID_MPEG_VIDEO_H263_MIN_QP

CID_MPEG_VIDEO_H263_P_FRAME_QP

CID_MPEG_VIDEO_H264_8X8_TRANSFORM

CID_MPEG_VIDEO_H264_ASO

CID_MPEG_VIDEO_H264_ASO_SLICE_ORDER

CID_MPEG_VIDEO_H264_B_FRAME_QP

CID_MPEG_VIDEO_H264_CPB_SIZE

CID_MPEG_VIDEO_H264_ENTROPY_MODE

CID_MPEG_VIDEO_H264_FMO

CID_MPEG_VIDEO_H264_FMO_CHANGE_DIRECTION

CID_MPEG_VIDEO_H264_FMO_CHANGE_RATE

CID_MPEG_VIDEO_H264_FMO_MAP_TYPE

CID_MPEG_VIDEO_H264_FMO_RUN_LENGTH

CID_MPEG_VIDEO_H264_FMO_SLICE_GROUP

CID_MPEG_VIDEO_H264_HIERARCHICAL_CODING

CID_MPEG_VIDEO_H264_HIERARCHICAL_CODING_LAYER

CID_MPEG_VIDEO_H264_HIERARCHICAL_CODING_LAYER_QP

CID_MPEG_VIDEO_H264_HIERARCHICAL_CODING_TYPE

CID_MPEG_VIDEO_H264_I_FRAME_QP

CID_MPEG_VIDEO_H264_I_PERIOD

CID_MPEG_VIDEO_H264_LEVEL

CID_MPEG_VIDEO_H264_LOOP_FILTER_ALPHA

CID_MPEG_VIDEO_H264_LOOP_FILTER_BETA

CID_MPEG_VIDEO_H264_LOOP_FILTER_MODE

CID_MPEG_VIDEO_H264_MAX_QP

CID_MPEG_VIDEO_H264_MIN_QP

CID_MPEG_VIDEO_H264_PROFILE

CID_MPEG_VIDEO_H264_P_FRAME_QP

CID_MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE

CID_MPEG_VIDEO_H264_SEI_FP_CURRENT_FRAME_0

CID_MPEG_VIDEO_H264_SEI_FRAME_PACKING

CID_MPEG_VIDEO_H264_VUI_EXT_SAR_HEIGHT

CID_MPEG_VIDEO_H264_VUI_EXT_SAR_WIDTH

CID_MPEG_VIDEO_H264_VUI_SAR_ENABLE

CID_MPEG_VIDEO_H264_VUI_SAR_IDC

CID_MPEG_VIDEO_HEADER_MODE

CID_MPEG_VIDEO_MAX_REF_PIC

CID_MPEG_VIDEO_MB_RC_ENABLE

CID_MPEG_VIDEO_MPEG4_B_FRAME_QP

CID_MPEG_VIDEO_MPEG4_I_FRAME_QP

CID_MPEG_VIDEO_MPEG4_LEVEL

CID_MPEG_VIDEO_MPEG4_MAX_QP

CID_MPEG_VIDEO_MPEG4_MIN_QP

CID_MPEG_VIDEO_MPEG4_PROFILE

CID_MPEG_VIDEO_MPEG4_P_FRAME_QP

CID_MPEG_VIDEO_MPEG4_QPEL

CID_MPEG_VIDEO_MULTI_SLICE_MAX_BYTES

CID_MPEG_VIDEO_MULTI_SLICE_MAX_MB

CID_MPEG_VIDEO_MULTI_SLICE_MODE

CID_MPEG_VIDEO_MUTE

CID_MPEG_VIDEO_MUTE_YUV

CID_MPEG_VIDEO_MV_H_SEARCH_RANGE

CID_MPEG_VIDEO_MV_V_SEARCH_RANGE

CID_MPEG_VIDEO_PULLDOWN

CID_MPEG_VIDEO_REPEAT_SEQ_HEADER

CID_MPEG_VIDEO_TEMPORAL_DECIMATION

CID_MPEG_VIDEO_VBV_DELAY

CID_MPEG_VIDEO_VBV_SIZE

CID_MPEG_VIDEO_VPX_1_PARTITION

CID_MPEG_VIDEO_VPX_1_REF_FRAME

CID_MPEG_VIDEO_VPX_2_PARTITIONS

CID_MPEG_VIDEO_VPX_2_REF_FRAME

CID_MPEG_VIDEO_VPX_3_REF_FRAME

CID_MPEG_VIDEO_VPX_4_PARTITIONS

CID_MPEG_VIDEO_VPX_8_PARTITIONS

CID_MPEG_VIDEO_VPX_FILTER_LEVEL

CID_MPEG_VIDEO_VPX_FILTER_SHARPNESS

CID_MPEG_VIDEO_VPX_GOLDEN_FRAME_REF_PERIOD

CID_MPEG_VIDEO_VPX_GOLDEN_FRAME_SEL

CID_MPEG_VIDEO_VPX_GOLDEN_FRAME_USE_PREV

CID_MPEG_VIDEO_VPX_GOLDEN_FRAME_USE_REF_PERIOD

CID_MPEG_VIDEO_VPX_IMD_DISABLE_4X4

CID_MPEG_VIDEO_VPX_I_FRAME_QP

CID_MPEG_VIDEO_VPX_MAX_QP

CID_MPEG_VIDEO_VPX_MIN_QP

CID_MPEG_VIDEO_VPX_NUM_PARTITIONS

CID_MPEG_VIDEO_VPX_NUM_REF_FRAMES

CID_MPEG_VIDEO_VPX_PROFILE

CID_MPEG_VIDEO_VPX_P_FRAME_QP

CID_PAN_ABSOLUTE

CID_PAN_RELATIVE

CID_PAN_RESET

CID_PAN_SPEED

CID_PILOT_TONE_DEVIATION

CID_PILOT_TONE_ENABLED

CID_PILOT_TONE_FREQUENCY

CID_PIXEL_RATE

CID_POWER_LINE_FREQUENCY

CID_POWER_LINE_FREQUENCY_50HZ

CID_POWER_LINE_FREQUENCY_60HZ

CID_POWER_LINE_FREQUENCY_AUTO

CID_POWER_LINE_FREQUENCY_DISABLED

CID_PRIVACY

CID_RDS_RECEPTION

CID_RDS_RX_MUSIC_SPEECH

CID_RDS_RX_PS_NAME

CID_RDS_RX_PTY

CID_RDS_RX_RADIO_TEXT

CID_RDS_RX_TRAFFIC_ANNOUNCEMENT

CID_RDS_RX_TRAFFIC_PROGRAM

CID_RDS_TX_ALT_FREQS

CID_RDS_TX_ALT_FREQS_ENABLE

CID_RDS_TX_ARTIFICIAL_HEAD

CID_RDS_TX_COMPRESSED

CID_RDS_TX_DEVIATION

CID_RDS_TX_DYNAMIC_PTY

CID_RDS_TX_MONO_STEREO

CID_RDS_TX_MUSIC_SPEECH

CID_RDS_TX_PI

CID_RDS_TX_PS_NAME

CID_RDS_TX_PTY

CID_RDS_TX_RADIO_TEXT

CID_RDS_TX_TRAFFIC_ANNOUNCEMENT

CID_RDS_TX_TRAFFIC_PROGRAM

CID_RED_BALANCE

CID_RF_TUNER_BANDWIDTH

CID_RF_TUNER_BANDWIDTH_AUTO

CID_RF_TUNER_CLASS

CID_RF_TUNER_CLASS_BASE

CID_RF_TUNER_IF_GAIN

CID_RF_TUNER_IF_GAIN_AUTO

CID_RF_TUNER_LNA_GAIN

CID_RF_TUNER_LNA_GAIN_AUTO

CID_RF_TUNER_MIXER_GAIN

CID_RF_TUNER_MIXER_GAIN_AUTO

CID_RF_TUNER_PLL_LOCK

CID_ROTATE

CID_SATURATION

CID_SCENE_MODE

CID_SHARPNESS

CID_TEST_PATTERN

CID_TEST_PATTERN_BLUE

CID_TEST_PATTERN_GREENB

CID_TEST_PATTERN_GREENR

CID_TEST_PATTERN_RED

CID_TILT_ABSOLUTE

CID_TILT_RELATIVE

CID_TILT_RESET

CID_TILT_SPEED

CID_TUNE_ANTENNA_CAPACITOR

CID_TUNE_DEEMPHASIS

CID_TUNE_POWER_LEVEL

CID_TUNE_PREEMPHASIS

CID_USER_ADV7180_BASE

CID_USER_BASE

CID_USER_BTTV_BASE

CID_USER_CLASS

CID_USER_MEYE_BASE

CID_USER_S2255_BASE

CID_USER_SAA7134_BASE

CID_USER_SI476X_BASE

CID_USER_TI_VPE_BASE

CID_VBLANK

CID_VFLIP

CID_WHITENESS

CID_WHITE_BALANCE_TEMPERATURE

CID_WIDE_DYNAMIC_RANGE

CID_ZOOM_ABSOLUTE

CID_ZOOM_CONTINUOUS

CID_ZOOM_RELATIVE

CLASS_CAMERA
Camera controls. Details.

CLASS_DETECT
Motion or object detection controls. Details.

CLASS_DV
Digital Video controls. Details.

CLASS_FLASH
Flash device controls. Details.

CLASS_FM_RX
FM Receiver controls. Details.

CLASS_FM_TX
FM Transmitter controls. Details.

CLASS_IMAGE_PROC
Image processing controls. Details.

CLASS_IMAGE_SOURCE
low-level controls of image source. Details.

CLASS_JPEG
JPEG compression controls. details.

CLASS_MPEG
MPEG compression controls. Details.

CLASS_RF_TUNER
RF tuner controls. Details.

CLASS_USER
User controls. Details.

COLORFX_ANTIQUE

COLORFX_AQUA

COLORFX_ART_FREEZE

COLORFX_BW

COLORFX_EMBOSS

COLORFX_GRASS_GREEN

COLORFX_NEGATIVE

COLORFX_NONE

COLORFX_SEPIA

COLORFX_SET_CBCR

COLORFX_SILHOUETTE

COLORFX_SKETCH

COLORFX_SKIN_WHITEN

COLORFX_SKY_BLUE

COLORFX_SOLARIZATION

COLORFX_VIVID

DEEMPHASIS_50_uS

DEEMPHASIS_75_uS

DEEMPHASIS_DISABLED

DETECT_MD_MODE_DISABLED

DETECT_MD_MODE_GLOBAL

DETECT_MD_MODE_REGION_GRID

DETECT_MD_MODE_THRESHOLD_GRID

DV_RGB_RANGE_AUTO

DV_RGB_RANGE_FULL

DV_RGB_RANGE_LIMITED

DV_TX_MODE_DVI_D

DV_TX_MODE_HDMI

EXPOSURE_APERTURE_PRIORITY

EXPOSURE_AUTO

EXPOSURE_MANUAL

EXPOSURE_METERING_AVERAGE

EXPOSURE_METERING_CENTER_WEIGHTED

EXPOSURE_METERING_MATRIX

EXPOSURE_METERING_SPOT

EXPOSURE_SHUTTER_PRIORITY

FIELD_ALTERNATE
Both fields alternating into separate buffers.

FIELD_ANY
None, top, bottom or interplaced depending on whatever it thinks is approximate.

FIELD_BOTTOM
Bottom field only.

FIELD_INTERLACED
Both fields interplaced.

FIELD_INTERLACED_BT
Both fields interplaced, top field first and the bottom field is transmitted first.

FIELD_INTERLACED_TB
Both fields interplaced, top field first and the top field is transmitted first.

FIELD_NONE
This device has no fields.

FIELD_SEQ_BT
Both fields sequential into one buffer, bottom-top order.

FIELD_SEQ_TB
Both fields sequential into one buffer, top-bottom order.

FIELD_TOP
Top field only.

FLAG_DISABLED
This control is permanently disabled and should be ignored by the application.

FLAG_EXECUTE_ON_WRITE
The value provided to the control will be propagated to the driver even if it remains constant. This is required when the control represents an action on the hardware. For example: clearing an error flag or triggering the flash.

FLAG_GRABBED
This control is temporarily unchangeable (e.g. another application controls resource).

FLAG_HAS_PAYLOAD
This control has a pointer type.

FLAG_INACTIVE
This control is not applicable to the current configuration.

FLAG_READ_ONLY
This control is permanently readable only.

FLAG_SLIDER
A hint that this control is best represented as a slider-like element in a user interface.

FLAG_UPDATE
Changing this control may affect the value of other controls within the same control class.

FLAG_VOLATILE
This control is volatile, which means that the value of the control changes continuously. A typical example would be the current gain value if the device is in auto-gain mode.

FLAG_WRITE_ONLY
This control is permanently writable only.

FLASH_FAULT_INDICATOR

FLASH_FAULT_INPUT_VOLTAGE

FLASH_FAULT_LED_OVER_TEMPERATURE

FLASH_FAULT_OVER_CURRENT

FLASH_FAULT_OVER_TEMPERATURE

FLASH_FAULT_OVER_VOLTAGE

FLASH_FAULT_SHORT_CIRCUIT

FLASH_FAULT_TIMEOUT

FLASH_FAULT_UNDER_VOLTAGE

FLASH_LED_MODE_FLASH

FLASH_LED_MODE_NONE

FLASH_LED_MODE_TORCH

FLASH_STROBE_SOURCE_EXTERNAL

FLASH_STROBE_SOURCE_SOFTWARE

ISO_SENSITIVITY_AUTO

ISO_SENSITIVITY_MANUAL

JPEG_ACTIVE_MARKER_APP0

JPEG_ACTIVE_MARKER_APP1

JPEG_ACTIVE_MARKER_COM

JPEG_ACTIVE_MARKER_DHT

JPEG_ACTIVE_MARKER_DQT

JPEG_CHROMA_SUBSAMPLING_410

JPEG_CHROMA_SUBSAMPLING_411

JPEG_CHROMA_SUBSAMPLING_420

JPEG_CHROMA_SUBSAMPLING_422

JPEG_CHROMA_SUBSAMPLING_444

JPEG_CHROMA_SUBSAMPLING_GRAY

LOCK_EXPOSURE

LOCK_FOCUS

LOCK_WHITE_BALANCE

MPEG_AUDIO_AC3_BITRATE_112K

MPEG_AUDIO_AC3_BITRATE_128K

MPEG_AUDIO_AC3_BITRATE_160K

MPEG_AUDIO_AC3_BITRATE_192K

MPEG_AUDIO_AC3_BITRATE_224K

MPEG_AUDIO_AC3_BITRATE_256K

MPEG_AUDIO_AC3_BITRATE_320K

MPEG_AUDIO_AC3_BITRATE_32K

MPEG_AUDIO_AC3_BITRATE_384K

MPEG_AUDIO_AC3_BITRATE_40K

MPEG_AUDIO_AC3_BITRATE_448K

MPEG_AUDIO_AC3_BITRATE_48K

MPEG_AUDIO_AC3_BITRATE_512K

MPEG_AUDIO_AC3_BITRATE_56K

MPEG_AUDIO_AC3_BITRATE_576K

MPEG_AUDIO_AC3_BITRATE_640K

MPEG_AUDIO_AC3_BITRATE_64K

MPEG_AUDIO_AC3_BITRATE_80K

MPEG_AUDIO_AC3_BITRATE_96K

MPEG_AUDIO_CRC_CRC16

MPEG_AUDIO_CRC_NONE

MPEG_AUDIO_DEC_PLAYBACK_AUTO

MPEG_AUDIO_DEC_PLAYBACK_LEFT

MPEG_AUDIO_DEC_PLAYBACK_MONO

MPEG_AUDIO_DEC_PLAYBACK_RIGHT

MPEG_AUDIO_DEC_PLAYBACK_STEREO

MPEG_AUDIO_DEC_PLAYBACK_SWAPPED_STEREO

MPEG_AUDIO_EMPHASIS_50_DIV_15_uS

MPEG_AUDIO_EMPHASIS_CCITT_J17

MPEG_AUDIO_EMPHASIS_NONE

MPEG_AUDIO_ENCODING_AAC

MPEG_AUDIO_ENCODING_AC3

MPEG_AUDIO_ENCODING_LAYER_1

MPEG_AUDIO_ENCODING_LAYER_2

MPEG_AUDIO_ENCODING_LAYER_3

MPEG_AUDIO_L1_BITRATE_128K

MPEG_AUDIO_L1_BITRATE_160K

MPEG_AUDIO_L1_BITRATE_192K

MPEG_AUDIO_L1_BITRATE_224K

MPEG_AUDIO_L1_BITRATE_256K

MPEG_AUDIO_L1_BITRATE_288K

MPEG_AUDIO_L1_BITRATE_320K

MPEG_AUDIO_L1_BITRATE_32K

MPEG_AUDIO_L1_BITRATE_352K

MPEG_AUDIO_L1_BITRATE_384K

MPEG_AUDIO_L1_BITRATE_416K

MPEG_AUDIO_L1_BITRATE_448K

MPEG_AUDIO_L1_BITRATE_64K

MPEG_AUDIO_L1_BITRATE_96K

MPEG_AUDIO_L2_BITRATE_112K

MPEG_AUDIO_L2_BITRATE_128K

MPEG_AUDIO_L2_BITRATE_160K

MPEG_AUDIO_L2_BITRATE_192K

MPEG_AUDIO_L2_BITRATE_224K

MPEG_AUDIO_L2_BITRATE_256K

MPEG_AUDIO_L2_BITRATE_320K

MPEG_AUDIO_L2_BITRATE_32K

MPEG_AUDIO_L2_BITRATE_384K

MPEG_AUDIO_L2_BITRATE_48K

MPEG_AUDIO_L2_BITRATE_56K

MPEG_AUDIO_L2_BITRATE_64K

MPEG_AUDIO_L2_BITRATE_80K

MPEG_AUDIO_L2_BITRATE_96K

MPEG_AUDIO_L3_BITRATE_112K

MPEG_AUDIO_L3_BITRATE_128K

MPEG_AUDIO_L3_BITRATE_160K

MPEG_AUDIO_L3_BITRATE_192K

MPEG_AUDIO_L3_BITRATE_224K

MPEG_AUDIO_L3_BITRATE_256K

MPEG_AUDIO_L3_BITRATE_320K

MPEG_AUDIO_L3_BITRATE_32K

MPEG_AUDIO_L3_BITRATE_40K

MPEG_AUDIO_L3_BITRATE_48K

MPEG_AUDIO_L3_BITRATE_56K

MPEG_AUDIO_L3_BITRATE_64K

MPEG_AUDIO_L3_BITRATE_80K

MPEG_AUDIO_L3_BITRATE_96K

MPEG_AUDIO_MODE_DUAL

MPEG_AUDIO_MODE_EXTENSION_BOUND_12

MPEG_AUDIO_MODE_EXTENSION_BOUND_16

MPEG_AUDIO_MODE_EXTENSION_BOUND_4

MPEG_AUDIO_MODE_EXTENSION_BOUND_8

MPEG_AUDIO_MODE_JOINT_STEREO

MPEG_AUDIO_MODE_MONO

MPEG_AUDIO_MODE_STEREO

MPEG_AUDIO_SAMPLING_FREQ_32000

MPEG_AUDIO_SAMPLING_FREQ_44100

MPEG_AUDIO_SAMPLING_FREQ_48000

MPEG_CX2341X_VIDEO_CHROMA_SPATIAL_FILTER_TYPE_1D_HOR

MPEG_CX2341X_VIDEO_CHROMA_SPATIAL_FILTER_TYPE_OFF

MPEG_CX2341X_VIDEO_LUMA_SPATIAL_FILTER_TYPE_1D_HOR

MPEG_CX2341X_VIDEO_LUMA_SPATIAL_FILTER_TYPE_1D_VERT

MPEG_CX2341X_VIDEO_LUMA_SPATIAL_FILTER_TYPE_2D_HV_SEPARABLE

MPEG_CX2341X_VIDEO_LUMA_SPATIAL_FILTER_TYPE_2D_SYM_NON_SEPARABLE

MPEG_CX2341X_VIDEO_LUMA_SPATIAL_FILTER_TYPE_OFF

MPEG_CX2341X_VIDEO_MEDIAN_FILTER_TYPE_DIAG

MPEG_CX2341X_VIDEO_MEDIAN_FILTER_TYPE_HOR

MPEG_CX2341X_VIDEO_MEDIAN_FILTER_TYPE_HOR_VERT

MPEG_CX2341X_VIDEO_MEDIAN_FILTER_TYPE_OFF

MPEG_CX2341X_VIDEO_MEDIAN_FILTER_TYPE_VERT

MPEG_CX2341X_VIDEO_SPATIAL_FILTER_MODE_AUTO

MPEG_CX2341X_VIDEO_SPATIAL_FILTER_MODE_MANUAL

MPEG_CX2341X_VIDEO_TEMPORAL_FILTER_MODE_AUTO

MPEG_CX2341X_VIDEO_TEMPORAL_FILTER_MODE_MANUAL

MPEG_MFC51_VIDEO_FORCE_FRAME_TYPE_DISABLED

MPEG_MFC51_VIDEO_FORCE_FRAME_TYPE_I_FRAME

MPEG_MFC51_VIDEO_FORCE_FRAME_TYPE_NOT_CODED

MPEG_MFC51_VIDEO_FRAME_SKIP_MODE_BUF_LIMIT

MPEG_MFC51_VIDEO_FRAME_SKIP_MODE_DISABLED

MPEG_MFC51_VIDEO_FRAME_SKIP_MODE_LEVEL_LIMIT

MPEG_STREAM_TYPE_MPEG1_SS

MPEG_STREAM_TYPE_MPEG1_VCD

MPEG_STREAM_TYPE_MPEG2_DVD

MPEG_STREAM_TYPE_MPEG2_PS

MPEG_STREAM_TYPE_MPEG2_SVCD

MPEG_STREAM_TYPE_MPEG2_TS

MPEG_STREAM_VBI_FMT_IVTV

MPEG_STREAM_VBI_FMT_NONE

MPEG_VIDEO_ASPECT_16x9

MPEG_VIDEO_ASPECT_1x1

MPEG_VIDEO_ASPECT_221x100

MPEG_VIDEO_ASPECT_4x3

MPEG_VIDEO_BITRATE_MODE_CBR

MPEG_VIDEO_BITRATE_MODE_VBR

MPEG_VIDEO_ENCODING_MPEG_1

MPEG_VIDEO_ENCODING_MPEG_2

MPEG_VIDEO_ENCODING_MPEG_4_AVC

MPEG_VIDEO_H264_ENTROPY_MODE_CABAC

MPEG_VIDEO_H264_ENTROPY_MODE_CAVLC

MPEG_VIDEO_H264_FMO_CHANGE_DIR_LEFT

MPEG_VIDEO_H264_FMO_CHANGE_DIR_RIGHT

MPEG_VIDEO_H264_FMO_MAP_TYPE_BOX_OUT

MPEG_VIDEO_H264_FMO_MAP_TYPE_EXPLICIT

MPEG_VIDEO_H264_FMO_MAP_TYPE_FOREGROUND_WITH_LEFT_OVER

MPEG_VIDEO_H264_FMO_MAP_TYPE_INTERLEAVED_SLICES

MPEG_VIDEO_H264_FMO_MAP_TYPE_RASTER_SCAN

MPEG_VIDEO_H264_FMO_MAP_TYPE_SCATTERED_SLICES

MPEG_VIDEO_H264_FMO_MAP_TYPE_WIPE_SCAN

MPEG_VIDEO_H264_HIERARCHICAL_CODING_B

MPEG_VIDEO_H264_HIERARCHICAL_CODING_P

MPEG_VIDEO_H264_LEVEL_1B

MPEG_VIDEO_H264_LEVEL_1_0

MPEG_VIDEO_H264_LEVEL_1_1

MPEG_VIDEO_H264_LEVEL_1_2

MPEG_VIDEO_H264_LEVEL_1_3

MPEG_VIDEO_H264_LEVEL_2_0

MPEG_VIDEO_H264_LEVEL_2_1

MPEG_VIDEO_H264_LEVEL_2_2

MPEG_VIDEO_H264_LEVEL_3_0

MPEG_VIDEO_H264_LEVEL_3_1

MPEG_VIDEO_H264_LEVEL_3_2

MPEG_VIDEO_H264_LEVEL_4_0

MPEG_VIDEO_H264_LEVEL_4_1

MPEG_VIDEO_H264_LEVEL_4_2

MPEG_VIDEO_H264_LEVEL_5_0

MPEG_VIDEO_H264_LEVEL_5_1

MPEG_VIDEO_H264_LOOP_FILTER_MODE_DISABLED

MPEG_VIDEO_H264_LOOP_FILTER_MODE_DISABLED_AT_SLICE_BOUNDARY

MPEG_VIDEO_H264_LOOP_FILTER_MODE_ENABLED

MPEG_VIDEO_H264_PROFILE_BASELINE

MPEG_VIDEO_H264_PROFILE_CAVLC_444_INTRA

MPEG_VIDEO_H264_PROFILE_CONSTRAINED_BASELINE

MPEG_VIDEO_H264_PROFILE_EXTENDED

MPEG_VIDEO_H264_PROFILE_HIGH

MPEG_VIDEO_H264_PROFILE_HIGH_10

MPEG_VIDEO_H264_PROFILE_HIGH_10_INTRA

MPEG_VIDEO_H264_PROFILE_HIGH_422

MPEG_VIDEO_H264_PROFILE_HIGH_422_INTRA

MPEG_VIDEO_H264_PROFILE_HIGH_444_INTRA

MPEG_VIDEO_H264_PROFILE_HIGH_444_PREDICTIVE

MPEG_VIDEO_H264_PROFILE_MAIN

MPEG_VIDEO_H264_PROFILE_MULTIVIEW_HIGH

MPEG_VIDEO_H264_PROFILE_SCALABLE_BASELINE

MPEG_VIDEO_H264_PROFILE_SCALABLE_HIGH

MPEG_VIDEO_H264_PROFILE_SCALABLE_HIGH_INTRA

MPEG_VIDEO_H264_PROFILE_STEREO_HIGH

MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE_CHECKERBOARD

MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE_COLUMN

MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE_ROW

MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE_SIDE_BY_SIDE

MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE_TEMPORAL

MPEG_VIDEO_H264_SEI_FP_ARRANGEMENT_TYPE_TOP_BOTTOM

MPEG_VIDEO_H264_VUI_SAR_IDC_10x11

MPEG_VIDEO_H264_VUI_SAR_IDC_12x11

MPEG_VIDEO_H264_VUI_SAR_IDC_15x11

MPEG_VIDEO_H264_VUI_SAR_IDC_160x99

MPEG_VIDEO_H264_VUI_SAR_IDC_16x11

MPEG_VIDEO_H264_VUI_SAR_IDC_18x11

MPEG_VIDEO_H264_VUI_SAR_IDC_1x1

MPEG_VIDEO_H264_VUI_SAR_IDC_20x11

MPEG_VIDEO_H264_VUI_SAR_IDC_24x11

MPEG_VIDEO_H264_VUI_SAR_IDC_2x1

MPEG_VIDEO_H264_VUI_SAR_IDC_32x11

MPEG_VIDEO_H264_VUI_SAR_IDC_3x2

MPEG_VIDEO_H264_VUI_SAR_IDC_40x33

MPEG_VIDEO_H264_VUI_SAR_IDC_4x3

MPEG_VIDEO_H264_VUI_SAR_IDC_64x33

MPEG_VIDEO_H264_VUI_SAR_IDC_80x33

MPEG_VIDEO_H264_VUI_SAR_IDC_EXTENDED

MPEG_VIDEO_H264_VUI_SAR_IDC_UNSPECIFIED

MPEG_VIDEO_HEADER_MODE_JOINED_WITH_1ST_FRAME

MPEG_VIDEO_HEADER_MODE_SEPARATE

MPEG_VIDEO_MPEG4_LEVEL_0

MPEG_VIDEO_MPEG4_LEVEL_0B

MPEG_VIDEO_MPEG4_LEVEL_1

MPEG_VIDEO_MPEG4_LEVEL_2

MPEG_VIDEO_MPEG4_LEVEL_3

MPEG_VIDEO_MPEG4_LEVEL_3B

MPEG_VIDEO_MPEG4_LEVEL_4

MPEG_VIDEO_MPEG4_LEVEL_5

MPEG_VIDEO_MPEG4_PROFILE_ADVANCED_CODING_EFFICIENCY

MPEG_VIDEO_MPEG4_PROFILE_ADVANCED_SIMPLE

MPEG_VIDEO_MPEG4_PROFILE_CORE

MPEG_VIDEO_MPEG4_PROFILE_SIMPLE

MPEG_VIDEO_MPEG4_PROFILE_SIMPLE_SCALABLE

MPEG_VIDEO_MULTI_SICE_MODE_MAX_BYTES

MPEG_VIDEO_MULTI_SICE_MODE_MAX_MB

MPEG_VIDEO_MULTI_SLICE_MODE_SINGLE

PREEMPHASIS_50_uS

PREEMPHASIS_75_uS

PREEMPHASIS_DISABLED

SCENE_MODE_BACKLIGHT

SCENE_MODE_BEACH_SNOW

SCENE_MODE_CANDLE_LIGHT

SCENE_MODE_DAWN_DUSK

SCENE_MODE_FALL_COLORS

SCENE_MODE_FIREWORKS

SCENE_MODE_LANDSCAPE

SCENE_MODE_NIGHT

SCENE_MODE_NONE

SCENE_MODE_PARTY_INDOOR

SCENE_MODE_PORTRAIT

SCENE_MODE_SPORTS

SCENE_MODE_SUNSET

SCENE_MODE_TEXT

WHITE_BALANCE_AUTO

WHITE_BALANCE_CLOUDY

WHITE_BALANCE_DAYLIGHT

WHITE_BALANCE_FLASH

WHITE_BALANCE_FLUORESCENT

WHITE_BALANCE_FLUORESCENT_H

WHITE_BALANCE_HORIZON

WHITE_BALANCE_INCANDESCENT

WHITE_BALANCE_MANUAL

WHITE_BALANCE_SHADE